Inline MyDateUtil into exercise 29
