+++ src/iii_conventions/n29OperatorsOverloading.kt
@@ -2,6 +2,7 @@ package iii_conventions
 
 import util.TODO
 import iii_conventions.TimeInterval.*
+import java.util.*
 
 fun todoTask29(): Nothing = TODO(
     """
@@ -29,3 +30,15 @@ fun task29_2(today: MyDate): MyDate {
 //    return today + YEAR * 2 + WEEK * 3 + DAY * 5
 }
 
+fun MyDate.nextDay() = addTimeIntervals(DAY, 1)
+
+fun MyDate.addTimeIntervals(timeInterval: TimeInterval, number: Int): MyDate {
+    val c = Calendar.getInstance()
+    c.set(year, month, dayOfMonth)
+    when (timeInterval) {
+        TimeInterval.DAY -> c.add(Calendar.DAY_OF_MONTH, number)
+        TimeInterval.WEEK -> c.add(Calendar.WEEK_OF_MONTH, number)
+        TimeInterval.YEAR -> c.add(Calendar.YEAR, number)
+    }
+    return MyDate(c.get(Calendar.YEAR), c.get(Calendar.MONTH), c.get(Calendar.DATE))
+}
